Render job refactoring
