--- src/main/java/com/sfengine/core/rendering/RenderJobFactory.java
@@ -1,7 +1,0 @@
-package com.sfengine.core.rendering;
-
-public class RenderJobFactory {
-
-
-
-}
